+++ backend/src/main/java/com/example/backend/repository/KategoriRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.repository;
 import com.example.backend.model.KategoriModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KategoriRepository extends JpaRepository<KategoriModel, Long> {
+public interface KategoriRepository extends JpaRepository<KategoriModel, Integer> {
 }
