+++ backend/src/main/java/com/example/backend/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserModel, Long> {
+public interface UserRepository extends JpaRepository<UserModel, Integer> {
 
     // Untuk validasi email saat register
     Optional<UserModel> findByEmail(String email);
